+++ src/components/tag/button/index.ts
@@ -1,0 +1,2 @@
+export { default as Button } from './Button';
+export { default as ButtonProps } from './ButtonProps';
