+++ src/components/tag/caption/CaptionProps.ts
