+++ src/components/tag/caption/index.ts
@@ -1,0 +1,2 @@
+export { default as Caption } from './Caption';
+export { default as CaptionProps } from './CaptionProps';
