+++ src/components/tag/image/index.ts
@@ -1,0 +1,2 @@
+export { default as Image } from './Image';
+export { default as ImageProps } from './ImageProps';
