--- src/components/tag/index.ts
@@ -1,8 +1,0 @@
-export { default as Button } from './Button';
-export { default as Caption } from './Caption';
-export { default as Image } from './Image';
-export { default as Input } from './Input';
-export { default as Label } from './Label';
-export { default as Option } from './Option';
-export { default as Select } from './Select';
-export { default as TextArea } from './TextArea';
