+++ src/components/tag/input/index.ts
@@ -1,0 +1,2 @@
+export { default as Input } from './Input';
+export { default as InputProps } from './InputProps';
