+++ src/components/tag/label/index.ts
@@ -1,0 +1,2 @@
+export { default as Label } from './Label';
+export { default as LabelProps } from './LabelProps';
