+++ src/components/tag/option/index.ts
@@ -1,0 +1,2 @@
+export { default as Option } from './Option';
+export { default as OptionProps } from './OptionProps';
