--- src/components/tag/props/index.ts
@@ -1,8 +1,0 @@
-export { default as ButtonProps } from './ButtonProps';
-export { default as CaptionProps } from './CaptionProps';
-export { default as ImageProps } from './ImageProps';
-export { default as InputProps } from './InputProps';
-export { default as LabelProps } from './LabelProps';
-export { default as OptionProps } from './OptionProps';
-export { default as SelectProps } from './SelectProps';
-export { default as TextAreaProps } from './TextAreaProps';
