+++ src/components/tag/select/index.ts
@@ -1,0 +1,2 @@
+export { default as Select } from './Select';
+export { default as SelectProps } from './SelectProps';
