+++ src/components/tag/textarea/index.ts
@@ -1,0 +1,2 @@
+export { default as TextArea } from './TextArea';
+export { default as TextAreaProps } from './TextAreaProps';
